requirements.txt
alembic==1.14.0
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
backoff==2.2.1
dnspython==2.7.0
greenlet==3.1.1
Mako==1.3.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
pymongo==4.10.1
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
